+++ tests/testthat/helper/runtt.R
@@ -100,7 +100,10 @@ local({
   )
   test.res <- if(packageVersion('testthat') > "1.0.2") {
     test_dir(
-      "testthat", reporter="check", env=environment(), filter=test.filt,
+      "testthat",
+      reporter="check",
+      env=environment(),
+      filter=test.filt,
       wrap=FALSE
     )
   } else {