+++ tests/testthat/testthat.demo.R
@@ -77,6 +77,7 @@ local({
   update_fastlm(.unitizer.fastlm, version="0.1.1")
   devtools::install(.unitizer.fastlm, quick=TRUE, quiet=TRUE, local=FALSE)
   unitizer:::read_line_set_vals(c("N", "N", "Y"))
+
   txt3 <- unitizer:::capture_output(
     untz3 <- unitize(.unitizer.test.file, interactive.mode=TRUE)
   )
@@ -125,8 +126,28 @@ local({
       "| No changes recorded.", fixed=TRUE
     )
   })
-  unitizer:::read_line_set_vals(c("Y", "Y", "Y"))
+  # Use this opportunity to make sure `use.diff=FALSE` works as intended
 
+  unitizer:::read_line_set_vals("Q")
+
+  txt5a <- unitizer:::capture_output(
+    unitize(.unitizer.test.file, interactive.mode=TRUE, use.diff=FALSE)
+  )
+  unitizer:::read_line_set_vals(c(".DIFF$state", "Q"))
+  txt5b <- unitizer:::capture_output(
+    unitize(.unitizer.test.file, interactive.mode=TRUE, use.diff=FALSE)
+  )
+  test_that("use.diff", {
+    expect_equal_to_reference(
+      txt5a, file.path("helper", "refobjs", "unitize_showdiff.rds")
+    )
+    expect_equal_to_reference(
+      txt5b, file.path("helper", "refobjs", "unitize_usediff_no.rds")
+    )
+  })
+  # Now actually accept the changes
+
+  unitizer:::read_line_set_vals(c("Y", "Y", "Y"))
   txt5 <- unitizer:::capture_output(
     untz5 <- unitize(.unitizer.test.file, interactive.mode=TRUE)
   )