+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## v1.4.5
 
+* Colors in test browser working again.
 * [#237](https://github.com/brodieG/unitizer/issues/237) Option to turn off
   diffs.
 * [#239](https://github.com/brodieG/unitizer/issues/239) Document issues with