+++ NEWS.md
@@ -3,6 +3,8 @@
 ## v1.4.5
 
 * Colors in test browser working again.
+* [#225](https://github.com/brodieG/unitizer/issues/225) Inferring file
+  locations when not in pkg top level.
 * [#237](https://github.com/brodieG/unitizer/issues/237) Option to turn off
   diffs.
 * [#239](https://github.com/brodieG/unitizer/issues/239) Document issues with